+++ MCFCRT/src/stdc/wchar/wmemcmp.c
@@ -15,8 +15,8 @@
 static inline uintptr_t wswap_ptr(uintptr_t w){
 #if __BYTE_ORDER__ == __ORDER_LITTLE_ENDIAN__
 	return _Generic(w,
-		uint32_t: __builtin_bswap32(((w << 8) & 0xFF00FF00        ) | ((w & 0xFF00FF00        ) >> 8)),
-		uint64_t: __builtin_bswap64(((w << 8) & 0xFF00FF00FF00FF00) | ((w & 0xFF00FF00FF00FF00) >> 8)));
+		uint32_t: __builtin_bswap32((uint32_t)(((w << 8) & 0xFF00FF00        ) | ((w & 0xFF00FF00        ) >> 8))),
+		uint64_t: __builtin_bswap64((uint64_t)(((w << 8) & 0xFF00FF00FF00FF00) | ((w & 0xFF00FF00FF00FF00) >> 8))));
 #else
 	return w;
 #endif
